requirements.txt
PyYAML == 5.1
aiohttp == 3.5.4
websockets == 4.0.1
